Move auth tests, add exception catching tests for ffmpeg read
